# specs/001-azure-devops-timetracker/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: SMIT TimeTrack API
  version: 0.1.0-draft
  description: >-
    Contrato inicial. O esquema de autenticação é provisório até conclusão da prova
    técnica documentada em research.md. Não implementar este bearer como token
    Azure DevOps sem validação de emissor, audiência e organização.
servers:
  - url: https://api.example.invalid/v1
security:
  - BackendSession: []
components:
  securitySchemes:
    BackendSession:
      type: http
      scheme: bearer
      description: Sessão curta emitida pelo backend após troca verificável de identidade; desenho pendente.
...
paths:
...
  /settings/organization:
    put:
      operationId: updateTimezone
      description: Fuso IANA da organização. Vale só para lançamentos novos; os antigos guardam data local e fuso próprios.
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [timezone], properties: { timezone: { type: string, example: America/Sao_Paulo } } }
      responses:
        '200': { description: Visão atualizada }
        '422': { description: Fuso inválido }
  /settings/policy:
    put:
      operationId: updatePolicy
      description: >-
        Nova versão das regras de lançamento, válida a partir de agora (as
        versões anteriores ficam guardadas). Incremento, duração mínima e
        comentário obrigatório valem para lançamento manual (a correção do
        administrador pelo relatório não passa por eles); limite diário e janela retroativa para
        qualquer lançamento/edição. Timer guarda segundos exatos.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [durationIncrementMinutes, dailyLimitHours, retroactiveWindowDays, commentRequired]
              properties:
                durationIncrementMinutes: { type: integer, enum: [1, 5, 10, 15, 30, 60] }
                minDurationMinutes: { type: integer, minimum: 1, maximum: 480, description: 'Duração mínima do lançamento manual; 1 = sem mínimo. Ausente mantém o valor em vigor.' }
                dailyLimitHours: { type: integer, minimum: 1, maximum: 24 }
                retroactiveWindowDays: { type: integer, minimum: 0, maximum: 365 }
                commentRequired: { type: boolean }
      responses:
        '200': { description: Visão atualizada }
        '422': { description: Valor inválido }
...
  /settings/overtime-rules:
    put:
      operationId: updateOvertimeRules
      description: >-
        Nova versão das regras de hora adicional, válida para os lançamentos
        feitos a partir de agora. Em dia útil, conta o que fica fora do
        expediente; sábado, domingo e feriado contam o dia todo. Cada parte é
        multiplicada pelo fator do dia; a parte noturna recebe ainda o adicional
        noturno e, se ligada, a hora reduzida (52min30s). Com
        `requireTimeOfDay`, lançamento manual sem início é recusado (422).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                enabled: { type: boolean }
                workdayStart: { type: string, example: '09:00' }
                workdayEnd: { type: string, example: '18:00', description: Depois do início. }
                factorWeekday: { type: number, minimum: 1, maximum: 5 }
                factorSaturday: { type: number, minimum: 1, maximum: 5 }
                factorSunday: { type: number, minimum: 1, maximum: 5 }
                factorHoliday: { type: number, minimum: 1, maximum: 5 }
                nightStart: { type: string, example: '22:00' }
                nightEnd: { type: string, example: '05:00' }
                nightPercent: { type: integer, minimum: 0, maximum: 100 }
                nightReducedHour: { type: boolean }
                requireTimeOfDay: { type: boolean }
                bankValidityMonths: { type: integer, minimum: 1, maximum: 12, description: Prazo para compensar o banco de horas. Omitido mantém o atual. }
                bankWeighted: { type: boolean, description: O banco recebe as horas ponderadas (com o fator). Omitido mantém o atual. }
                alertDailyExtraHours: { type: number, minimum: 0, maximum: 12, description: 'Aviso: horas extras por dia útil (padrão 2; 0 desliga). Omitido mantém o atual.' }
                alertWeeklyHours: { type: number, minimum: 0, maximum: 99, description: 'Aviso: horas por semana (padrão 44; 0 desliga).' }
                alertRestHours: { type: number, minimum: 0, maximum: 24, description: 'Aviso: descanso mínimo entre jornadas (padrão 11; 0 desliga).' }
      responses:
        '200': { description: Visão atualizada (campo `overtime`) }
        '422': { description: Valor inválido }
  /settings/holidays:
    post:
      operationId: addHoliday
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [date, name]
              properties:
                date: { type: string, format: date }
                name: { type: string, maxLength: 100 }
      responses:
        '200': { description: Visão atualizada (campo `holidays`) }
        '422': { description: Data já cadastrada ou inválida }
  /settings/holidays/national:
    post:
      operationId: addNationalHolidays
      description: Acrescenta os feriados nacionais do ano (inclui Sexta-feira Santa) que ainda não estão no calendário.
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [year], properties: { year: { type: integer } } }
      responses:
        '200': { description: Visão atualizada }
  /settings/holidays/{holidayId}:
    delete:
      operationId: removeHoliday
      parameters:
        - { name: holidayId, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Visão atualizada }
  /settings/members/{memberId}/hours-regime:
    put:
      operationId: setHoursRegime
      description: clt (hora extra, banco ou a pagar), pj (só a pagar) ou none (não controla jornada; não gera hora adicional).
      parameters:
        - { name: memberId, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [regime], properties: { regime: { type: string, enum: [clt, pj, none] } } }
      responses:
        '200': { description: Visão atualizada }
  /settings/members/{memberId}/overtime-profile:
    put:
      operationId: setOvertimeProfile
      description: >-
        Perfil de hora extra (só tem efeito para CLT): preapproved (hora extra
        sem pedido e sem aviso), standard (sem pedido aprovado, fica sujeita à
        aprovação) ou restricted (o trecho fora do expediente sem pedido
        aprovado vira hora extra a confirmar). Auditado.
      parameters:
        - { name: memberId, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [profile], properties: { profile: { type: string, enum: [preapproved, standard, restricted] } } }
      responses:
        '200': { description: 'Visão atualizada (members[].overtimeProfile)' }
  /settings/projects/{projectId}:
    patch:
      operationId: setProjectEnabled
      description: Projeto desabilitado recusa novos lançamentos e timers; o histórico permanece.
      parameters:
        - { name: projectId, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, properties: { enabled: { type: boolean }, countsAsIdle: { type: boolean, description: 'No relatório por projeto, as horas entram em Horas Ociosas.' }, usesBillable: { type: boolean, description: 'O projeto cobra o cliente por hora; desligado (padrão), nenhuma hora dele é faturável e o campo some do lançamento.' } }, description: 'Ao menos um dos campos.' }
      responses:
        '200': { description: Visão atualizada }
        '404': { description: Projeto de outra organização ou inexistente }
  /settings/activity-types:
    post:
      operationId: createActivityType
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, maxLength: 100 }
                color: { type: [string, 'null'], pattern: '^#[0-9A-Fa-f]{6}$' }
                defaultBillable: { type: boolean }
      responses:
        '200': { description: Visão atualizada }
        '422': { description: Nome repetido ou cor inválida }
  /settings/activity-types/{typeId}:
    patch:
      operationId: updateActivityType
      description: Renomear, recolorir, faturável padrão e habilitar/desabilitar (não há exclusão; o histórico mantém o tipo).
      parameters:
        - { name: typeId, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, maxLength: 100 }
                color: { type: [string, 'null'] }
                enabled: { type: boolean }
                defaultBillable: { type: boolean }
      responses:
        '200': { description: Visão atualizada }
        '404': { description: Tipo de outra organização ou inexistente }
        '422': { description: Nome repetido ou cor inválida }
  /settings/people/sync:
    post:
      operationId: syncPeople
      description: >-
        Atualiza a lista de pessoas com quem tem licença ativa no Azure DevOps
        (lida pela extensão com o acesso do administrador). Cria quem ainda não
        abriu a extensão, sem nenhum papel (sem acesso até o administrador
        liberar), atualiza nomes e marca como inativo (`directoryActive: false`)
        quem saiu da lista; nunca apaga nem altera papéis ou lançamentos. A
        identidade é comparada sem diferenciar maiúsculas. Lista vazia é recusada.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [people]
              properties:
                people:
                  type: array
                  minItems: 1
                  maxItems: 5000
                  items:
                    type: object
                    required: [identityId, displayName]
                    properties:
                      identityId: { type: string, format: uuid }
                      displayName: { type: string, maxLength: 200 }
      responses:
        '200': { description: Visão atualizada (members com directoryActive e peopleSyncedAt) }
        '422': { description: Lista vazia ou identidade inválida }
  /settings/role-assignments:
    post:
      operationId: grantRole
      description: Concede papel (member, approver, manager, admin) na organização inteira ou num projeto. Idempotente.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [memberId, role]
              properties:
                memberId: { type: string }
                role: { type: string, enum: [member, approver, manager, admin] }
                projectId: { type: [string, 'null'] }
      responses:
        '200': { description: Visão atualizada }
        '404': { description: Pessoa ou projeto de outra organização }
        '422': { description: Papel inválido }
  /settings/role-assignments/{assignmentId}:
    delete:
      operationId: revokeRole
      parameters:
        - { name: assignmentId, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Visão atualizada }
        '404': { description: Atribuição de outra organização }
        '409': { description: Seria removido o último administrador }
  /settings/approver-assignments:
    post:
      operationId: designateApprover
      description: >-
        `approverId` aprova as semanas de `memberId` (em todos os projetos ou num
        só). Resolvido no envio; `applyToPending` também reatribui as semanas
        já enviadas e pendentes dessa pessoa (decididas nunca mudam).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [memberId, approverId]
              properties:
                memberId: { type: string }
                approverId: { type: string }
                projectId: { type: [string, 'null'] }
                applyToPending: { type: boolean }
      responses:
        '200': { description: Visão atualizada }
        '404': { description: Pessoa ou projeto de outra organização }
        '422': { description: Aprovador igual à própria pessoa }
  /settings/approver-assignments/{assignmentId}:
    delete:
      operationId: removeDesignation
      parameters:
        - { name: assignmentId, in: path, required: true, schema: { type: string } }
      requestBody:
        content:
          application/json:
            schema: { type: object, properties: { applyToPending: { type: boolean } } }
      responses:
        '200': { description: Visão atualizada }
        '404': { description: Designação de outra organização }
...